# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: PuneKar MVP API
  description: Lightweight civic app to report, prioritise, verify and track ward-level public issues.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        firebase_uid:
          type: string
        email:
          type: string
        display_name:
          type: string
        photo_url:
          type: string
        ward:
          type: string
    Issue:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [pothole, garbage, streetlight, encroachment]
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        image_url:
          type: string
        status:
          type: string
          enum: [reported, in_progress, resolved, rejected]
        upvotes:
          type: integer
        created_by:
          type: integer
        created_at:
          type: string
          format: date-time
    DashboardSummary:
      type: object
      properties:
        total_issues:
          type: integer
        resolved_issues:
          type: integer
        pending_issues:
          type: integer
        categories:
          type: object
          additionalProperties:
            type: integer
paths:
  /api/auth/verify:
    post:
      summary: Verify Firebase Token & Login/Signup user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
  /api/users/me:
    get:
      summary: Get current authenticated user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/issues:
    get:
      summary: List issues for a ward
      parameters:
        - name: ward
          in: query
          required: false
          schema:
            type: string
            default: Viman Nagar
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [top, new]
            default: new
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Issue'
    post:
      summary: Create a new issue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, category, latitude, longitude, image_url]
              properties:
                title:
                  type: string
                description:
                  type: string
                category:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
                image_url:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
  /api/issues/{id}:
    get:
      summary: Get issue details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '404':
          description: Not Found
    patch:
      summary: Change issue status (Admin/Mod)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
      responses:
        '200':
          description: OK
  /api/issues/{id}/verify:
    post:
      summary: Upload verification (after-photo)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [photo_url]
              properties:
                photo_url:
                  type: string
                note:
                  type: string
      responses:
        '201':
          description: Created
  /api/issues/{id}/vote:
    post:
      summary: Upvote an issue
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    delete:
      summary: Remove upvote from an issue
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/dashboard/summary:
    get:
      summary: Dashboard summary counts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'